+++ test/views/github.js
@@ -3,13 +3,13 @@ import { preconnect } from '@aegisjsproject/router/router.js';
 import { url } from '@aegisjsproject/url/url.js';
 
 const [users, setUsers] = manageState('github:users', {});
-// const url = (strings, ...values) => String.raw(strings, ...values.map(val => encodeURIComponent(val)));
 
 preconnect('https://api.github.com');
 preconnect('https://avatars.githubusercontent.com');
 
 export default async ({ matches, signal }) => {
 	const username = matches?.pathname?.groups?.username;
+
 	if (typeof username !== 'string') {
 		return `<p>Error: No GitHub username provided.</p>`;
 	} else if (! (username in users)) {
@@ -25,18 +25,19 @@ export default async ({ matches, signal }) => {
 
 	signal.addEventListener('abort', ({ target }) => console.info(target.reason), { once: true });
 
-	return `
-		<section>
-			<template shadowrootmode="open">
-				<h1>${users[username].name ?? username}</h1>
-				<p><strong>Username:</strong> ${users[username].login}</p>
-				<p><strong>Bio:</strong> ${users[username].bio ?? 'No bio available'}</p>
-				<p><strong>Public Repos:</strong> ${users[username].public_repos}</p>
-				<p><strong>Followers:</strong> ${users[username].followers}</p>
-				<p><strong>Following:</strong> ${users[username].following}</p>
-				${users[username].blog ? `<p><a href="${users[username].blog}" target="_blank">Website</a></p>` : ''}
-				${users[username].avatar_url ? `<img src="${users[username].avatar_url}" alt="Avatar" crossorigin="anonymous" referrerpolicy="no-referrer" loading="lazy" width="100" />` : ''}
-			</template>
-		</section>
-	`;
+	return `<section>
+		<template shadowrootmode="open">
+			<h1>${users[username].name ?? username}</h1>
+			<p><strong>Username:</strong> ${users[username].login}</p>
+			<p><strong>Bio:</strong> ${users[username].bio ?? 'No bio available'}</p>
+			<p><strong>Public Repos:</strong> ${users[username].public_repos}</p>
+			<p><strong>Followers:</strong> ${users[username].followers}</p>
+			<p><strong>Following:</strong> ${users[username].following}</p>
+			${users[username].blog ? `<p><a href="${users[username].blog}" target="_blank">Website</a></p>` : ''}
+			${users[username].avatar_url ? `<img src="${users[username].avatar_url}" alt="Avatar" crossorigin="anonymous" referrerpolicy="no-referrer" loading="lazy" width="100" />` : ''}
+		</template>
+	</section>`;
 };
+
+export const title = ({ matches }) => `GitHub Profile for ${matches.pathname.groups.username}`;
+export const description = ({ matches }) => `GitHub Profile for ${matches.pathname.groups.username}`;